remove initRealm(). There is no going back!
