+++ lib/src/realm_flutter.dart
@@ -16,7 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////////
 
-export 'realm_class.dart' hide RealmInternal, Results, Helpers, DynamicObject;
+export 'realm_class.dart' hide RealmInternal;
 
 /// Is Realm loaded in Flutter or Dart
 const bool isFlutterPlatform = true;
