+++ test/realm_test.dart
@@ -68,8 +68,6 @@ void main([List<String>? args]) {
 
   print("Current PID $pid");
 
-  initRealm();
-
   setUp(() async {
     var currentDir = Directory.current;
     if (Platform.isAndroid || Platform.isIOS) {
